Refactor file list and tree creation
The objective was to reduce memory usage by discarding the structs we get from
Drive, keeping only a minimal set.  Initial benchmarking indicates it doesn't
make any difference.  It may be that the memory isn't free'd quite as
agressively as I expect by Go.

This also paves the way for periodically refreshing the file list.