Add a version to the Checkpoint, and db.reinit()
The Version field of the checkpoint allows a check for an increase.  If
found at setup time, throw away all of the previous Metadata entries
from Google Drive, and reset the Changes counter to resync.

This paves the way for changing what's stored in leveldb that won't
result in insanity across versions.