Add flag for inodeCacheSize, default from 1k->10k
Add a flag to allow you to specify the size of the drive_db's inode
cache.  The default was a constant of 1000.  Directories containing
more files than will fit in the inode cache suffer a large performance
penalty (~13x).  The cache is relatively cheap, memory-wise, so I
also changed the default to 10k entries.